initial sparkResource implementation
